refactor: Split llm_d runtime helpers
